refactor: Mentions and entities modules
