--- docs/reference/bot/api/commands.md
@@ -1,2 +1,0 @@
-::: botx.bot.api.commands.bot_disabled_response
-::: botx.bot.api.commands.accepted_response
