--- docs/reference/bot/handler_collector.md
@@ -1,1 +1,0 @@
-::: botx.bot.handler_collector.HandlerCollector
