--- docs/reference/bot/system_events.md
@@ -1,12 +1,0 @@
-## `chat_created`
-::: botx.bot.models.commands.system_events.chat_created
-
----
-
-## `deleted_from_chat`
-::: botx.bot.models.commands.system_events.deleted_from_chat
-
----
-
-## `added_to_chat`
-::: botx.bot.models.commands.system_events.added_to_chat
