--- docs/reference/models.md
@@ -1,27 +1,0 @@
-
-## `ChatTypes`
-::: botx.shared_models.chat_types.ChatTypes
-    handler: python
-    selection:
-      members:
-
-
-## `ChatListItem`
-::: botx.client.chats_api.list_chats.ChatListItem
-    handler: python
-    selection:
-      members:
-
-
-## `ChatInfo`
-::: botx.client.chats_api.chat_info.ChatInfo
-    handler: python
-    selection:
-      members:
-
-
-## `UserFromSearch`
-::: botx.client.users_api.user_from_search.UserFromSearch
-    handler: python
-    selection:
-      members:
